👌 Notable events, entity enum, formatting

+++ src/main/resources/db/migration/V1.1.0__notification.sql
@@ -2,6 +2,6 @@
 CREATE TABLE notification (
     id BIGINT PRIMARY KEY,
     message TEXT,
-    type_id BIGINT,
+    type_id INT,
     user_id BIGINT
 );
